+++ src/router/index.ts
@@ -1,0 +1,32 @@
+export { defineRoutes } from "./defineRoutes";
+export { createRouter } from "./createRouter";
+export { redirect, RedirectError } from "./types";
+export type {
+  RouteDefinition,
+  RouterOptions,
+  MeshRouter,
+  RouteMatch,
+  NavigationTarget,
+  NavigationContext,
+  RouteMiddleware,
+  RouteGuard,
+  RouteLoader,
+  RouteBeforeLoad,
+  RouteComponentLoader,
+  RouteDependencyFn,
+  RouteErrorRecovery,
+  KeepAliveOptions,
+  AnalyticsOptions,
+  RouteAnalyticsEvent,
+  PredictivePrefetchOptions,
+  OfflineOptions,
+  RouterState,
+  RouterContextValue,
+  RouteHistoryEntry
+} from "./types";
+export { compilePattern, matchPattern, buildPath, matchRoutes, flattenRoutes, interpolatePath, normalizePath, parseSearch, serializeSearch } from "./matchRoutes";
+export type { FlatRoute } from "./matchRoutes";
+export { createBrowserHistory, createMemoryHistory } from "./historyAdapter";
+export type { HistoryAdapter, HistoryLocation, HistoryEntry } from "./historyAdapter";
+export { updateDocumentMeta } from "./meta";
+export type { RouteMeta } from "./meta";